Add a "daily" mode, for running via cronjob or similar
Daily mode changes the start date of processing, to be the current
time period and the one immediately before it.

eg yesterday and today, last week and this week, last month and
   this month

Its purpose is for running via cronjob (or similar) at least once
a day, to update the stats for the most recent time periods.

It does the previous time period as well as the current one, to
ensure all of the entries for the previous time period have been
processed.

In theory, this should be suitable for running much more often
than daily.  eg every minute should work.  Not sure if that's a
good idea or not though. ;)